fixed "indexPage" api method

+++ src/main/java/searchengine/services/IndexingService.java
@@ -6,16 +6,18 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import searchengine.config.Site;
 import searchengine.config.SitesList;
-import searchengine.dao.*;
-import searchengine.dto.statistics.*;
+import searchengine.dao.IndexDao;
+import searchengine.dao.LemmaDao;
+import searchengine.dao.PageDao;
+import searchengine.dao.SiteDao;
+import searchengine.dto.statistics.ErrorResponse;
+import searchengine.dto.statistics.Response;
 import searchengine.model.Page;
 import searchengine.model.SiteStatus;
 
-import java.io.IOException;
+import javax.persistence.NoResultException;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ForkJoinPool;
@@ -33,7 +35,7 @@ public class IndexingService {
     private final LemmaDao lemmaRepository;
     private final IndexDao indexRepository;
 
-    public void startIndexing() throws IOException {
+    public void startIndexing(){
         try {
             MappingService.clearData();
         } catch (Exception e){
@@ -50,9 +52,17 @@ public class IndexingService {
 
         ExecutorService executor = Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors());
 
-        for (Site site : sitesList) {
+        for (Site configSite : sitesList) {
             executor.submit(() -> {
-                mapper(site.getUrl());
+                String url = configSite.getUrl();
+                String siteName = url.replace("https://", "").replace("http://", "").replace("/", "");
+                searchengine.model.Site site = new searchengine.model.Site();
+                site.setName(siteName);
+                site.setLastError(null);
+                site.setStatus(SiteStatus.INDEXING);
+                site.setStatusTime(LocalDateTime.now());
+                site.setUrl(url);
+                mapper(url, site);
             });
         }
         executor.shutdown();
@@ -60,15 +70,8 @@ public class IndexingService {
 
 
 
-    private void mapper(String url){
-        String siteName = url.replace("https://", "").replace("http://", "").replace("/", "");
-        searchengine.model.Site site = new searchengine.model.Site();
-        site.setName(siteName);
-        site.setLastError(null);
-        site.setStatus(SiteStatus.INDEXING);
-        site.setStatusTime(LocalDateTime.now());
-        site.setUrl(url);
-        siteRepository.save(site);
+    private void mapper(String url, searchengine.model.Site site){
+        siteRepository.saveOrUpdate(site);
         MappingService service = new MappingService(url, pageRepository, site, siteRepository, lemmaRepository, indexRepository);
         new ForkJoinPool().invoke(service);
         site = service.getSite();
@@ -78,24 +81,30 @@ public class IndexingService {
         if (inProcess) {
             site.setStatus(SiteStatus.INDEXED);
             siteRepository.update(site);
-        } else {
-            site.setStatus(SiteStatus.FAILED);
-            site.setLastError("Индексация прервана пользователем");
-            siteRepository.update(site);
         }
-    }
-
-
-
-    public static void stop() {
         inProcess = false;
         MappingService.setIsStopped(true);
     }
 
-    public static void process(){
+
+    public void stop() {
+        inProcess = false;
+        MappingService.setIsStopped(true);
+        siteRepository.findAll().forEach(site -> {
+            if (site.getStatus().equals(SiteStatus.INDEXING)){
+                site.setStatus(SiteStatus.FAILED);
+                site.setLastError("Индексация прервана пользователем");
+                siteRepository.update(site);
+            }
+        });
+    }
+
+    public void process(){
         inProcess = true;
         MappingService.setIsStopped(false);
     }
+
+
 
     public Object indexPage(String url){
         try {
@@ -104,6 +113,15 @@ public class IndexingService {
 
         }
         process();
+        boolean pageIsInSite = false;
+        for(Site site : sites.getSites()){
+            if (url.startsWith(site.getUrl())){
+                pageIsInSite = true;
+            }
+        }
+        if (!pageIsInSite) {
+            return new ErrorResponse("Данная страница находится за пределами сайтов, указанных в конфигурационном файле");
+        }
         String path = url.replace("https://", "").replace("http://", "");
         if (url.charAt(url.length() - 1) != '/'){
             path = path.substring(path.indexOf('/'));
@@ -111,40 +129,37 @@ public class IndexingService {
             path = StringUtils.chop(path.substring(path.indexOf('/')));
         }
         String mainUrl = url.substring(0, url.indexOf('/', 8));
-        searchengine.model.Site site = (searchengine.model.Site) siteRepository.findByUrl(mainUrl);
-        List<Page> pages = pageRepository.findBySiteId(site.getId());
-        int countDeleted = 0;
-        for (Page page : pages){
-            if (page.getPath().startsWith(path)){
-                try {
-                    pageRepository.delete(page);
-                    countDeleted++;
-                } catch (Exception e){
 
+        searchengine.model.Site site;
+
+        try {
+            site = (searchengine.model.Site) siteRepository.findByUrl(mainUrl);
+        } catch (NoResultException e){
+            String siteName = mainUrl.replace("https://", "").replace("http://", "").replace("/", "");
+            site = new searchengine.model.Site();
+            site.setStatus(SiteStatus.INDEXING);
+            site.setName(siteName);
+            site.setStatusTime(LocalDateTime.now());
+            site.setUrl(mainUrl);
+            site.setLastError(null);
+        }
+        List<Page> pages = pageRepository.findBySiteId(site.getId());
+        if (!pages.isEmpty()) {
+            for (Page page : pages) {
+                if (page.getPath().startsWith(path)) {
+                    try {
+                        pageRepository.delete(page);
+                    } catch (Exception e){}
                 }
             }
         }
-        if (countDeleted == 0){
-            ErrorResponse response = new ErrorResponse("Данная страница находится за пределами сайтов, указанных в конфигурационном файле");
-            return response;
-        }
-        Response response = new Response();
-
+        searchengine.model.Site currentSite = site;
         ExecutorService executor = Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors());
         executor.submit(() -> {
-            new ForkJoinPool().invoke(new MappingService(url, pageRepository, site, siteRepository, lemmaRepository, indexRepository));
+            mapper(url, currentSite);
         });
         executor.shutdown();
-
-        if (inProcess) {
-            site.setStatus(SiteStatus.INDEXED);
-            siteRepository.update(site);
-        } else {
-            site.setStatus(SiteStatus.FAILED);
-            site.setLastError("Индексация прервана пользователем");
-            siteRepository.update(site);
-        }
-        return response;
+        return new Response();
     }
 
 }
